Graphics.Gloss.Accelerate.Data.Picture

Description

Documentation

bitmapOfArray :: Array DIM2 Word32 -> Bool -> Picture

Use an Accelerate array of RGBA data as a bitmap image. If the image is generated programatically every frame, then the second parameter should be False. If you have loaded it from static data then use True.